requirements.txt
accelerate
huggingface-hub
ipython
matplotlib
mediapipe
ml-dtypes
numpy
openai
opencv-contrib-python
packaging
pandas
pillow
requests
safetensors
scipy
tokenizers
torch
torchvision
tqdm
transformers
diffusers
